Move getNormalizedPath() into separate header
