+++ src/common/strutil.h
@@ -56,13 +56,6 @@ bool contains(const std::string &str, const std::string &s);
  * \return a vector of split strings
  */
 std::vector<std::string> split(const std::string &str, const std::regex &split);
-
-/*!
- * An utility function that normalizes file path by:
- * - replacing \ with /
- * - lowercasing the path
- */
-std::string getNormalizedPath(const std::string &path);
 }
 
 #endif //AWE_STRUTIL_H
